segunda versao com jwt criada e funcionando

+++ app/controllers/delete_user_controller.py
@@ -1,22 +1,22 @@
 from flask_jwt_extended import jwt_required, get_jwt_identity
-from flask_httpauth import HTTPTokenAuth
 from app.models.user_model import UserModel
 from flask import current_app, session
+from app.exc import UserNotFound
 
-auth = HTTPTokenAuth(scheme='Bearer')
+@jwt_required()
+def delete_user_controller():
 
-@auth.verify_token
-def verify_token(token):
-    find_token = UserModel.query.filter_by(api_key=token).first()
-    return find_token   
-
-@auth.login_required
-def delete_user_controller(id):
     session = current_app.db.session
-    user = auth.current_user()
+    current_user = get_jwt_identity()
 
-    session.delete(user)
-    session.commit()
-    return {"msg": f"user {user.name} has been deleted."},204
-
+    try:
+        user = UserModel.query.filter_by(email=current_user["email"]).first()
+        if not user:
+            raise UserNotFound
+        session.delete(user)
+        session.commit()
+        return {"msg": f"user {user.name} has been deleted."},204
+    
+    except UserNotFound as err:
+        return err.message, 404
